+++ lib/adventure.rb
@@ -4,6 +4,7 @@ require 'adventure/version'
 
 require 'adventure/actor'
 require 'adventure/building'
+require 'adventure/dungeon'
 require 'adventure/quest'
 require 'adventure/settlement'